kill old modal
