+++ src/components/CurrencyInputPanel/currency-panel.scss
@@ -158,15 +158,13 @@
 }
 
 .token-modal {
+  @extend %col-nowrap;
   background-color: $white;
-  position: relative;
-  bottom: 21rem;
   width: 100%;
-  height: 21rem;
-  z-index: 2000;
+  height: 100%;
+
   border-top-left-radius: 1rem;
   border-top-right-radius: 1rem;
-  transition: 250ms ease-in-out;
 
   &__search-container {
     @extend %row-nowrap;
@@ -183,8 +181,8 @@
   }
 
   &__token-list {
-    -webkit-overflow-scrolling: touch;
-    height: 17.5rem;
+    flex-grow: 1;
+    height: 100%;
     overflow-y: auto;
   }
 
@@ -256,21 +254,6 @@
 
   &__token-label {
     color: $dove-gray;
-  }
-
-  @media only screen and (min-width: 768px) {
-    max-width: 560px;
-    max-height: 768px;
-    position: absolute;
-    margin-left: auto;
-    margin-right: auto;
-    border-radius: 1rem;
-    padding-bottom: 1rem;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    top: 0;
-    margin-top: 4rem;
   }
 }
