--- src/components/Modal/modal.scss
@@ -1,18 +1,0 @@
-@import '../../variables.scss';
-
-.modal-container {
-  position: relative;
-  height: 100vh;
-  width: 100vw;
-  background-color: rgba($black, 0.6);
-  z-index: 1000;
-}
-
-.modal-container-appear {
-  opacity: 0.01;
-}
-
-.modal-container-appear.modal-container-appear-active {
-  opacity: 1;
-  transition: opacity 200ms ease-in-out;
-}
