--- src/components/QrCode/qr-code.scss
@@ -1,15 +1,0 @@
-.qr-code {
-  &__video {
-    height: 100%;
-    width: 100%;
-  }
-
-  &__modal {
-    z-index: 2000;
-    position: absolute;
-    top: 100px;
-    bottom: 100px;
-    right: 100px;
-    left: 100px;
-  }
-}
