+++ src/constants/index.js
@@ -1,4 +1,6 @@
 export const FACTORY_ADDRESSES = {
   1: '0xc0a47dFe034B400B47bDaD5FecDa2621de6c4d95',
-  4: '0xf5D915570BC477f9B8D6C0E980aA81757A3AaC36'
+  3: '0x9c83dCE8CA20E9aAF9D3efc003b2ea62aBC08351',
+  4: '0xf5D915570BC477f9B8D6C0E980aA81757A3AaC36',
+  42: '0xD3E51Ef092B2845f10401a0159B2B96e8B6c3D30'
 }
