+++ src/pages/Pool/ModeSelector.js
@@ -1,7 +1,6 @@
 import React, { useState, useCallback } from 'react'
 import { withRouter, NavLink } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
-import { CSSTransitionGroup } from 'react-transition-group'
 
 import OversizedPanel from '../../components/OversizedPanel'
 import Dropdown from '../../assets/images/dropdown-blue.svg'
@@ -31,7 +30,7 @@ const poolTabOrder = [
 function ModeSelector({ location: { pathname }, history }) {
   const { t } = useTranslation()
 
-  const [isShowingModal, setIsShowingModal] = useState(false)
+  const [modalIsOpen, setModalIsOpen] = useState(false)
 
   const activeTabKey = poolTabOrder[poolTabOrder.findIndex(({ regex }) => pathname.match(regex))].textKey
 
@@ -49,53 +48,44 @@ function ModeSelector({ location: { pathname }, history }) {
     navigate(-1)
   }, [navigate])
 
-  useBodyKeyDown('ArrowDown', navigateRight, isShowingModal)
-  useBodyKeyDown('ArrowUp', navigateLeft, isShowingModal)
+  useBodyKeyDown('ArrowDown', navigateRight, modalIsOpen)
+  useBodyKeyDown('ArrowUp', navigateLeft, modalIsOpen)
 
   return (
     <OversizedPanel hideTop>
       <div
         className="pool__liquidity-container"
         onClick={() => {
-          setIsShowingModal(true)
+          setModalIsOpen(true)
         }}
       >
         <span className="pool__liquidity-label">{t(activeTabKey)}</span>
         <img src={Dropdown} alt="dropdown" />
       </div>
-      {isShowingModal && (
-        <Modal
-          onClose={() => {
-            setIsShowingModal(false)
-          }}
-        >
-          <CSSTransitionGroup
-            transitionName="pool-modal"
-            transitionAppear={true}
-            transitionLeave={true}
-            transitionAppearTimeout={200}
-            transitionLeaveTimeout={200}
-            transitionEnterTimeout={200}
-          >
-            <div className="pool-modal">
-              {poolTabOrder.map(({ path, textKey, regex }) => (
-                <NavLink
-                  key={path}
-                  to={path}
-                  className="pool-modal__item"
-                  activeClassName="pool-modal__item--selected"
-                  isActive={(_, { pathname }) => pathname.match(regex)}
-                  onClick={() => {
-                    setIsShowingModal(false)
-                  }}
-                >
-                  {t(textKey)}
-                </NavLink>
-              ))}
-            </div>
-          </CSSTransitionGroup>
-        </Modal>
-      )}
+      <Modal
+        isOpen={modalIsOpen}
+        onDismiss={() => {
+          setModalIsOpen(false)
+        }}
+        minHeight={null}
+      >
+        <div className="pool-modal">
+          {poolTabOrder.map(({ path, textKey, regex }) => (
+            <NavLink
+              key={path}
+              to={path}
+              className="pool-modal__item"
+              activeClassName="pool-modal__item--selected"
+              isActive={(_, { pathname }) => pathname.match(regex)}
+              onClick={() => {
+                setModalIsOpen(false)
+              }}
+            >
+              {t(textKey)}
+            </NavLink>
+          ))}
+        </div>
+      </Modal>
     </OversizedPanel>
   )
 }
