+++ src/pages/Pool/pool.scss
@@ -97,29 +97,9 @@
 
 .pool-modal {
   background-color: $white;
-  position: relative;
-  bottom: 11.0625rem;
   width: 100%;
-  height: 11.0625rem;
-  z-index: 2000;
-  border-top-left-radius: 1rem;
-  border-top-right-radius: 1rem;
-  transition: 250ms ease-in-out;
-  padding: 1rem 0 0.5rem;
-
-  @media only screen and (min-width: 768px) {
-    max-width: 560px;
-    position: absolute;
-    margin-left: auto;
-    margin-right: auto;
-    border-radius: 1rem;
-    padding-bottom: 1rem;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    top: 0;
-    margin-top: 4rem;
-  }
+  height: 100%;
+  padding: 2rem 0 2rem 0;
 
   &__item {
     @extend %row-nowrap;
@@ -140,14 +120,6 @@
       color: $royal-blue;
     }
   }
-}
-
-.pool-modal-appear {
-  bottom: 0;
-}
-
-.pool-modal-appear.modal-container-appear-active {
-  bottom: 0;
 }
 
 .create-exchange {
